Move uid, fibonacciDelaySequence, and sleep into lib/utils.js
